Semeter-4/LensLink/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
cloudinary==1.44.1
Django==6.0.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==12.1.1
psycopg2-binary>=2.9.0
PyJWT==2.11.0
python-dotenv==1.2.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3<3.0.0
gunicorn==24.0.0
whitenoise==6.9.0
dj-database-url==2.3.0
